# server/swagger.yml
swagger: '2.0'
info:
  description: API for Events Mananger Application
  version: 1.0.0
  title: Events Manager
  # Contact info for your development or API team
  contact:
    email: emmanuel.nonye.abaye@gmail.com

# tags are used for organizing operations
...
paths:
...
  /api/v1/centers:
    get:
      tags:
      - developers
      - centers
      summary: Get details of all centers
      operationId: get-all-centers
      description: Gets the details of all centers
      produces:
      - application/json
      responses:
        200:
          description: Success, Centers found
          schema:
            type: object
            properties:
              status:
                type: string
              message:
                type: string
              data:
                type: array
                items:
                  $ref: '#/definitions/Center'
          examples:
            application/json:
              status: 'Success'
              message: 'Centers  found'
              data:
                - id: '1'
                  name: 'City Hall'
                  description: 'Best center for your indoor events'
                  location: 'Lagos'
                  capacity: '1000'
                  facilities: 'Car park, Bathroom/Toilet'
                  price: '200000'
                  userId: '1'
                  image: '#noImage'
                - id: '2'
                  name: 'Event Manager Park'
                  description: 'Best center for your outdoor events'
                  location: 'Lagos'
                  capacity: '1000'
                  facilities: 'Car park, Bathroom/Toilet'
                  price: '200000'
                  userId: '1'
                  image: '#noImage' 
        500:
          description: 'Center not found'
          schema:
            type: object
            properties:
              status:
                type: string
              message:
                type: string
          examples:
            application/json:
              status: 'Error'
              message: 'Server Error'
    post:
      tags:
      - admins
      - centers
      summary: Create an event center
      operationId: createCenter
      description: Creates a new center for an admin
      consumes:
      - multipart/form-data
      produces:
      - application/json
      parameters:
      - in: header
        name: x-access-token
        type: string
        description: authorization header
        required: true
      - in: formData
        name: name
        type: string
        description: Name of center
        required: true
      - in: formData
        name: description
        type: string
        description: Description of center
        required: true
      - in: formData
        name: location
        type: string
        description: Location of center
        required: true
      - in: formData
        name: capacity
        type: string
        description: Number of persons for a center
        required: true
      - in: formData
        name: facilities
        type: string
        description: Facilities of event center
        required: true
      - in: formData
        name: price
        type: string
        description: Price of center
        required: true
      - in: formData
        name: image
        type: file
        description: Image showing event center
        required: false
      responses:
        200:
          description: Successfully created an center
          schema:
            type: object
            properties:
              status:
                type: string
              message:
                type: string
              data:
                $ref: '#/definitions/Center'
          examples:
            application/json:
              status: 'Success'
              message: 'Center has been  created'
              data:
                id: '1'
                name: 'City Hall'
                description: 'Best center for your indoor events'
                location: 'Lagos'
                capacity: '1000'
                facilities: 'Car park, Bathroom/Toilet'
                price: '200000'
                userId: '1'
                image: '#noImage'
        400:
          description: Invalid input
          schema:
            type: object
            properties:
              message:
                type: string
          examples:
            application/json:
              message: "Detail is required"
        500:
          description: Server Error
          schema:
            type: object
            properties:
              status:
                type: string
              message:
                type: string
          examples:
            application/json:
              status: 'Error'
              message: 'There was an error in adding center pls try again'
  /api/v1/centers/{centerId}:
    get:
      tags:
      - admins
      - centers
      summary: Get details of a center
      operationId: get-a-center
      description: Gets the details of a center using its id
      produces:
      - application/json
      parameters:
      - in: path
        name: centerId
        type: integer
        minimum: 1
        description:  Id of event center
        required: true
      responses:
        200:
          description: Success, Center found
          schema:
            type: object
            properties:
              status:
                type: string
              message:
                type: string
              data:
                $ref: '#/definitions/Center'
          examples:
            application/json:
              status: 'Success'
              message: 'Center found'
              data:
                id: '1'
                name: 'City Hall'
                description: 'Best center for your indoor events'
                location: 'Lagos'
                capacity: '1000'
                facilities: 'Car park, Bathroom/Toilet'
                price: '200000'
                userId: '1'
                image: '#noImage'
        400:
          description: Invalid input
          schema:
            type: object
            properties:
              message:
                type: string
          examples:
            application/json:
              message: "Detail is required"
        404:
          description: 'Center not found'
          schema:
            type: object
            properties:
              status:
                type: string
              message:
                type: string
          examples:
            application/json:
              status: 'Error'
              message: 'Center not found'
        500:
          description: 'Server error'
          schema:
            type: object
            properties:
              status:
                type: string
              message:
                type: string
          examples:
            application/json:
              status: 'Error'
              message: 'Server Error'
    put:
      tags:
      - admins
      - centers
      summary: Update an event center
      operationId: update-center
      description: Updates an existing center for an admin
      consumes:
      - multipart/form-data
      produces:
      - application/json
      parameters:
      - in: header
        name: x-access-token
        type: string
        description: authorization header
        required: true
      - in: path
        name: centerId
        type: integer
        minimum: 1
        description: Id of center
        required: true
      - in: formData
        name: name
        type: string
        description: Name of center
        required: true
      - in: formData
        name: description
        type: string
        description: Description of center
        required: true
      - in: formData
        name: location
        type: string
        description: Location of center
        required: true
      - in: formData
        name: capacity
        type: string
        description: Number of persons for a center
        required: true
      - in: formData
        name: facilities
        type: string
        description: Facilities of event center
        required: true
      - in: formData
        name: price
        type: string
        description: Price of center
        required: true
      - in: formData
        name: available
        type: string
        description: Availablity of center
        required: true
      - in: formData
        name: image
        type: file
        description: Image showing event center
        required: false
      responses:
        200:
          description: Successfully updated a center
          schema:
            type: object
            properties:
              status:
                type: string
              message:
                type: string
              data:
                $ref: '#/definitions/Center'
          examples:
            application/json:
              status: 'Success'
              message: 'Center Updated'
              data: [1]
        400:
          description: Invalid input
          schema:
            type: object
            properties:
              message:
                type: string
          examples:
            application/json:
              message: "Detail is required"
        403:
          description: Forbidden from updating center
          schema:
            type: object
            properties:
              message:
                type: string
          examples:
            application/json:
              message: "You do not have privilege to modify this Center"
        500:
          description: Server Error
          schema:
            type: object
            properties:
              status:
                type: string
              message:
                type: string
          examples:
            application/json:
              status: 'Error'
              message: 'Cannot update center due to server error'
    delete:
      tags:
      - admins
      - centers
      summary: Delete an event center
      operationId: delete-center
      description: Deletes an existing center for an admin
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - in: path
        name: centerId
        type: integer
        minimum: 1
        description: id of the center admin wants to delete
        required: true
      - in: header
        name: x-access-token
        type: string
        description: authorization header
        required: true
      responses:
        200:
          description: Successfully deleted a center
          schema:
            type: object
            properties:
              status:
                type: string
              message:
                type: string
              data:
                type: string
          examples:
            application/json:
              status: 'Success'
              message: 'Center has been  deleted'
              data: '[1]'
        400:
          description: Invalid input
          schema:
            type: object
            properties:
              message:
                type: string
          examples:
            application/json:
              message: "Detail is required"
        403:
          description: Forbidden from deleting center
          schema:
            type: object
            properties:
              message:
                type: string
          examples:
            application/json:
              message: "You do not have privilege to delete this Center"
        500:
          description: Server Error
          schema:
            type: object
            properties:
              status:
                type: string
              message:
                type: string
          examples:
            application/json:
              status: 'Error'
              message: 'There was an error in deleting the center. Please try again later.' 
definitions:
...
  Center:
    type: object
    properties:
      id:
        type: string
      name:
        type: string
      description:
        type: string
      location:
        type: string
      capacity:
        type: string
      facilities:
        type: string
      price:
        type: string
      userId:
        type: string
      image:
        type: string